# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Client Management Api",
    "description": "API doc for Client Management Project"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Clients",
      "description": "Clients API"
    },
    {
      "name": "Providers",
      "description": "Providers API"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/client": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create a client",
        "parameters": [
          {
            "name": "client",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createClient"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get all clients",
        "description": "Returns a clients list",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getClients"
            }
          }
        }
      }
    },
    "/client/{clientId}": {
      "description": "Returns a single client",
      "parameters": [
        {
          "name": "clientID",
          "in": "path",
          "required": true,
          "description": "ID of client to return",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Find client by ID",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "404": {
            "description": "Client not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Find client by ID and update",
        "description": "Update a client",
        "parameters": [
          {
            "name": "client",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/updateClient"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "delete": {
        "summary": "Find client by ID and delete",
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/deleteResponse"
            }
          },
          "404": {
            "description": "Client not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/providers": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get all providers",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getProviders"
            }
          }
        }
      }
    },
    "/provider": {
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Create a new provider",
        "parameters": [
          {
            "name": "provider",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/provider"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getProvider"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/provider/{providerId}": {
      "parameters": [
        {
          "name": "providerId",
          "in": "path",
          "required": true,
          "description": "The provider ID",
          "type": "string"
        }
      ],
      "put": {
        "tags": [
          "Providers"
        ],
        "summary": "Find provider by ID and update",
        "parameters": [
          {
            "name": "provider",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/provider"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/getProvider"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "404": {
            "description": "Provider not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      },
      "delete": {
        "summary": "Find provider by ID and delete",
        "tags": [
          "Providers"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/deleteResponse"
            }
          },
          "404": {
            "description": "Provider not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "createClient": {
      "required": [
        "name",
        "email",
        "phone"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "updateClient": {
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "getClient": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "getClients": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/getClient"
      }
    },
    "provider": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "getProvider": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "getProviders": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/getProvider"
      }
    },
    "deleteResponse": {
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "errorResponse": {
      "properties": {
        "error": {
          "properties": {
            "message": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "private": {
              "type": "boolean"
            },
            "status": {
              "type": "number"
            }
          }
        }
      }
    }
  }
}